requirements.txt
fastapi
pydantic
pymongo
python-dotenv
uvicorn[standard]